+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'chat-ai'
+rootProject.name = 'chat-app'
