--- src/main/java/com/li/chatai/ChatAiApplication.java
@@ -1,13 +1,0 @@
-package com.li.chatai;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ChatAiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ChatAiApplication.class, args);
-    }
-
-}
