--- src/main/java/com/li/chatai/domain/chat/chatMessage/dto/ChatMessageDto.java
@@ -1,5 +1,0 @@
-package com.li.chatai.domain.chat.chatMessage.dto;
-
-
-public class ChatMessageDto {
-}
