--- src/main/java/com/li/chatai/domain/chat/chatMessage/repository/ChatMessageRepository.java
@@ -1,12 +1,0 @@
-package com.li.chatai.domain.chat.chatMessage.repository;
-
-import com.li.chatai.domain.chat.chatMessage.entity.ChatMessage;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
-    List<ChatMessage> getAllById(Long id);
-}
