--- src/main/java/com/li/chatai/domain/chat/chatRoom/dto/RequestCreateRoom.java
@@ -1,10 +1,0 @@
-package com.li.chatai.domain.chat.chatRoom.dto;
-
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class RequestCreateRoom {
-    private String roomName;
-}
