--- src/main/java/com/li/chatai/domain/chat/chatRoom/repository/ChatRoomRepository.java
@@ -1,9 +1,0 @@
-package com.li.chatai.domain.chat.chatRoom.repository;
-
-import com.li.chatai.domain.chat.chatRoom.entity.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-}
