+++ src/main/java/com/li/chatapp/ChatAiApplication.java
@@ -1,0 +1,13 @@
+package com.li.chatapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ChatAiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ChatAiApplication.class, args);
+    }
+
+}
