+++ src/main/java/com/li/chatapp/domain/chat/chatMessage/dto/ChatMessageDto.java
@@ -1,0 +1,5 @@
+package com.li.chatapp.domain.chat.chatMessage.dto;
+
+
+public class ChatMessageDto {
+}
