+++ src/main/java/com/li/chatapp/domain/chat/chatMessage/repository/ChatMessageRepository.java
@@ -1,0 +1,12 @@
+package com.li.chatapp.domain.chat.chatMessage.repository;
+
+import com.li.chatapp.domain.chat.chatMessage.entity.ChatMessage;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+    List<ChatMessage> getAllById(Long id);
+}
