+++ src/main/java/com/li/chatapp/domain/chat/chatRoom/dto/RequestCreateRoom.java
@@ -1,0 +1,10 @@
+package com.li.chatapp.domain.chat.chatRoom.dto;
+
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class RequestCreateRoom {
+    private String roomName;
+}
