+++ src/main/java/com/li/chatapp/domain/chat/chatRoom/repository/ChatRoomRepository.java
@@ -1,0 +1,9 @@
+package com.li.chatapp.domain.chat.chatRoom.repository;
+
+import com.li.chatapp.domain.chat.chatRoom.entity.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
+}
