+++ src/main/java/com/li/chatapp/domain/global/jpa/BaseEntity.java
@@ -1,0 +1,31 @@
+package com.li.chatapp.domain.global.jpa;
+
+import groovy.transform.EqualsAndHashCode;
+import groovy.transform.ToString;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Getter
+@SuperBuilder
+@MappedSuperclass
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+@ToString
+@EqualsAndHashCode
+public class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @CreatedDate
+    private Long createdAt;
+    @LastModifiedDate
+    private Long updatedAt;
+}
