+++ src/test/java/com/li/chatai/member/member/service/MemberServiceTest.java
@@ -1,0 +1,4 @@
+package com.li.chatai.member.member.service;
+
+public class MemberServiceTest {
+}
